Better structure of provider classes
