+++ .vscode/settings.json
@@ -45,16 +45,51 @@
 	"files.trimTrailingWhitespace": true,
 	"files.insertFinalNewline": true,
 	"files.exclude": {
+		"**/.git": true,
+		"**/.svn": true,
+		"**/.hg": true,
+		"**/.DS_Store": true,
+		"**/Thumbs.db": true,
 		".git": true,
 		".build": true,
 		".profile-oss": true,
-		"**/.DS_Store": true,
 		".vscode-test": true,
 		"cli/target": true,
 		"build/**/*.js.map": true,
 		"build/**/*.js": {
 			"when": "$(basename).ts"
-		}
+		},
+		"[!e]*/**": true,
+		"e[!x]*/**": true,
+		"ex[!t]*/**": true,
+		"ext[!e]*/**": true,
+		"exte[!n]*/**": true,
+		"exten[!s]*/**": true,
+		"extens[!i]*/**": true,
+		"extensi[!o]*/**": true,
+		"extensio[!n]*/**": true,
+		"extension[!s]*/**": true,
+		"extensions/[!p]*/**": true,
+		"extensions/p[!o]*/**": true,
+		"extensions/po[!s]*/**": true,
+		"extensions/pos[!i]*/**": true,
+		"extensions/posi[!t]*/**": true,
+		"extensions/posit[!r]*/**": true,
+		"extensions/positr[!o]*/**": true,
+		"extensions/positro[!n]*/**": true,
+		"extensions/positron[!-]*/**": true,
+		"extensions/positron-[!a]*/**": true,
+		"extensions/positron-a[!s]*/**": true,
+		"extensions/positron-as[!s]*/**": true,
+		"extensions/positron-ass[!i]*/**": true,
+		"extensions/positron-assi[!s]*/**": true,
+		"extensions/positron-assis[!t]*/**": true,
+		"extensions/positron-assist[!a]*/**": true,
+		"extensions/positron-assista[!n]*/**": true,
+		"extensions/positron-assistan[!t]*/**": true,
+		"extensions/positron-assistant/[!s]*/**": true,
+		"extensions/positron-assistant/s[!r]*/**": true,
+		"extensions/positron-assistant/sr[!c]*/**": true
 	},
 	"files.associations": {
 		"cglicenses.json": "jsonc",
