+++ extensions/positron-assistant/src/providers/base/modelProvider.ts
@@ -6,74 +6,64 @@
 import * as vscode from 'vscode';
 import * as positron from 'positron';
 import * as ai from 'ai';
-import { ModelConfig, SecretStorage, getMaxConnectionAttempts, getProviderTimeoutMs } from '../../config';
-import { processMessages, toAIMessage, isAuthorizationError } from '../../utils';
+import { ModelConfig, SecretStorage, getMaxConnectionAttempts } from '../../config';
+import { isAuthorizationError } from '../../utils';
 import { applyModelFilters } from '../../modelFilters';
-import { TokenUsage } from '../../tokens';
 import { getAllModelDefinitions } from '../../modelDefinitions';
-import { createModelInfo, getMaxTokens, markDefaultModel } from '../../modelResolutionHelpers';
+import { createModelInfo, markDefaultModel } from '../../modelResolutionHelpers';
 import { DEFAULT_MAX_TOKEN_INPUT, DEFAULT_MAX_TOKEN_OUTPUT } from '../../constants';
-import { log, recordRequestTokenUsage, recordTokenUsage } from '../../extension';
 import { ModelProviderLogger } from './modelProviderLogger';
 import { AuthenticationError, ModelRetrievalError } from './modelProviderErrors';
-
-/**
- * Result of provider autoconfiguration attempt.
- *
- * @deprecated Use {@link AutoconfigureResult} from modelProviderTypes.ts instead.
- */
-export interface AutoconfigureResult {
-	/**
-	 * Whether the provider was successfully configured.
-	 */
-	configured: boolean;
-
-	/**
-	 * Optional message describing the configuration result or any issues encountered.
-	 */
-	message?: string;
-
-	/**
-	 * Configuration values that were set during autoconfiguration.
-	 * For example: { apiKey: string, baseUrl: string }
-	 */
-	configuration?: Record<string, any>;
-}
+import { AutoconfigureResult, AIProviderFactory } from './modelProviderTypes';
 
 /**
  * Abstract base class for all model providers in the Positron Assistant extension.
  *
- * This class provides a unified interface for interacting with various AI model providers
+ * This class provides a universal interface for interacting with various AI model providers
  * (Anthropic, OpenAI, Google, Azure, etc.) through a consistent API. It implements the
  * {@link positron.ai.LanguageModelChatProvider} interface to integrate with Positron's
  * language model system.
+ *
+ * This base class is provider-agnostic and can support ANY provider implementation:
+ * - Vercel AI SDK providers (via {@link VercelModelProvider})
+ * - Custom SDK providers (direct implementation)
+ * - Hybrid providers (custom auth + Vercel chat)
  *
  * Key responsibilities:
  * - Connection resolution with retry logic and timeout handling
  * - Model discovery and filtering based on capabilities and configuration
  * - Credential validation and authentication error handling
- * - Token counting and usage tracking
- * - Chat response streaming with progress reporting
- * - Tool/function calling support
+ * - Token counting interface
  * - Provider-specific error parsing
  *
- * Providers can be implemented in two ways:
- * 1. **Vercel AI SDK providers** (default): Use the Vercel AI SDK for model interactions
- * 2. **Custom providers**: Implement custom logic for providers not supported by Vercel AI SDK
+ * Subclasses must implement:
+ * - {@link provideLanguageModelChatResponse} - Chat response generation
+ * - {@link sendTestMessage} - Connection testing
+ * - {@link createAIProvider} - Provider initialization
  *
  * @example
  * ```typescript
- * class MyProvider extends ModelProvider {
+ * class MyCustomProvider extends ModelProvider {
  *   get providerName(): string {
  *     return 'My Provider';
  *   }
  *
  *   protected createAIProvider() {
- *     return createMyProvider({ apiKey: this._config.apiKey });
+ *     // Initialize your custom SDK
+ *     return undefined;
+ *   }
+ *
+ *   async provideLanguageModelChatResponse(...) {
+ *     // Implement custom chat logic
+ *   }
+ *
+ *   protected async sendTestMessage(modelId: string) {
+ *     // Implement custom test logic
  *   }
  * }
  * ```
  *
+ * @see {@link VercelModelProvider} for Vercel AI SDK providers
  * @see {@link ModelProviderLogger} for logging functionality
  * @see {@link ModelProviderErrors} for error types
  */
@@ -98,32 +88,6 @@ export abstract class ModelProvider implements positron.ai.LanguageModelChatProv
 	 * Unique identifier for this model configuration.
 	 */
 	public readonly id: string;
-
-	/**
-	 * The AI provider instance for Vercel AI SDK-based providers.
-	 * This function creates a language model instance given a model ID and optional configuration.
-	 * Optional to support custom (non-Vercel SDK) provider implementations.
-	 */
-	protected aiProvider?: (id: string, options?: Record<string, any>) => ai.LanguageModelV1;
-
-	/**
-	 * Custom provider implementation for non-Vercel SDK providers.
-	 * Use this when the provider doesn't have a Vercel AI SDK integration.
-	 */
-	protected customProvider?: any;
-
-	/**
-	 * Type of provider implementation being used.
-	 * - 'vercel': Uses Vercel AI SDK (default)
-	 * - 'custom': Uses custom implementation
-	 */
-	protected providerType: 'vercel' | 'custom' = 'vercel';
-
-	/**
-	 * Additional options passed to the AI provider when creating model instances.
-	 * Provider-specific options like temperature, top_p, etc.
-	 */
-	protected aiOptions: Record<string, any> = {};
 
 	/**
 	 * Cached list of available models for this provider.
@@ -161,16 +125,39 @@ export abstract class ModelProvider implements positron.ai.LanguageModelChatProv
 		this.id = _config.id;
 		this.name = _config.name;
 		this.provider = _config.provider;
-		// Logger initialization deferred to subclass since providerName is abstract
-		this.logger = null as any; // Will be initialized by subclass calling initializeLogger()
+		// Logger will be initialized by subclass
+		this.logger = null as any;
+		// Call initialization template method to set up logger and provider
+		this.initialize();
 	}
 
 	/**
-	 * Initializes the logger. Must be called by subclass constructor after providerName is available.
+	 * Template method for provider initialization.
+	 * Calls initializeLogger() and initializeProvider() in the correct order.
+	 * This is called automatically during construction.
+	 * @private
+	 */
+	private initialize(): void {
+		this.initializeLogger();
+		this.initializeProvider();
+	}
+
+	/**
+	 * Initializes the logger. Called automatically during construction.
 	 * @protected
 	 */
 	protected initializeLogger(): void {
 		this.logger = new ModelProviderLogger(this.providerName);
+	}
+
+	/**
+	 * Optional provider-specific initialization hook.
+	 * Override this method in subclasses to set up the AI provider instance.
+	 * Called automatically during construction after logger initialization.
+	 * @protected
+	 */
+	protected initializeProvider(): void {
+		// Default implementation does nothing - subclasses override as needed
 	}
 
 	/**
@@ -194,28 +181,27 @@ export abstract class ModelProvider implements positron.ai.LanguageModelChatProv
 	 * Creates the AI provider instance for this provider.
 	 *
 	 * This method is called during provider initialization to set up the underlying
-	 * AI SDK provider (e.g., Anthropic, OpenAI). For Vercel AI SDK-based providers,
-	 * return the provider factory function. For custom providers, return undefined
-	 * and implement {@link provideCustomResponse} instead.
+	 * AI SDK provider. The return value depends on the provider type:
+	 * - Vercel AI SDK providers: Return the provider factory function
+	 * - Custom providers: Return undefined and initialize custom SDK in this method
 	 *
 	 * @returns The AI provider factory function, or undefined for custom providers
 	 *
 	 * @example
 	 * ```typescript
-	 * // Vercel AI SDK provider
+	 * // Vercel AI SDK provider (use VercelModelProvider instead)
 	 * protected createAIProvider() {
 	 *   return createAnthropic({ apiKey: this._config.apiKey });
 	 * }
 	 *
 	 * // Custom provider
 	 * protected createAIProvider() {
-	 *   this.customProvider = new MyCustomProvider();
-	 *   this.providerType = 'custom';
+	 *   this._customClient = new MyCustomProvider();
 	 *   return undefined;
 	 * }
 	 * ```
 	 */
-	protected abstract createAIProvider(): ((id: string, options?: Record<string, any>) => ai.LanguageModelV1) | undefined;
+	protected abstract createAIProvider(): AIProviderFactory | undefined;
 
 	/**
 	 * Validates the provider's credentials before attempting to connect.
@@ -376,36 +362,35 @@ export abstract class ModelProvider implements positron.ai.LanguageModelChatProv
 	/**
 	 * Sends a test message to verify that a specific model is accessible and responsive.
 	 *
-	 * For Vercel AI SDK providers, this uses {@link ai.generateText} with a simple
-	 * prompt and includes a timeout and retry logic. Custom providers must override
-	 * this method to implement their own test logic.
-	 *
-	 * The test message is kept simple ("I'm checking to see if you're there...") to
-	 * minimize token usage while still verifying end-to-end connectivity.
+	 * Subclasses must implement this method to test connectivity with their specific
+	 * provider SDK. The test message should be kept simple to minimize token usage
+	 * while still verifying end-to-end connectivity.
 	 *
 	 * @param modelId - The ID of the model to test
 	 * @returns A promise that resolves to the test response from the model
 	 *
-	 * @throws {Error} If the provider type is not configured
-	 * @throws {Error} If custom provider test is not implemented (default behavior)
-	 * @throws {ai.AISDKError} If the model request fails
+	 * @throws {Error} If the model request fails
+	 *
+	 * @example
+	 * ```typescript
+	 * // Vercel AI SDK provider
+	 * protected async sendTestMessage(modelId: string) {
+	 *   return ai.generateText({
+	 *     model: this.aiProvider(modelId),
+	 *     prompt: "Hello",
+	 *     maxRetries: 1,
+	 *   });
+	 * }
+	 *
+	 * // Custom provider
+	 * protected async sendTestMessage(modelId: string) {
+	 *   return this._client.chat({ model: modelId, messages: [{ role: 'user', content: 'Hello' }] });
+	 * }
+	 * ```
 	 *
 	 * @see {@link getProviderTimeoutMs} for timeout configuration
 	 */
-	protected async sendTestMessage(modelId: string): Promise<any> {
-		if (this.providerType === 'vercel' && this.aiProvider) {
-			return ai.generateText({
-				model: this.aiProvider(modelId, this.aiOptions),
-				prompt: `I'm checking to see if you're there. Respond only with the word "hello".`,
-				abortSignal: AbortSignal.timeout(getProviderTimeoutMs()),
-				maxRetries: 1, // Retry the request once in case of transient errors
-			});
-		} else if (this.customProvider) {
-			// Handle custom provider test - override in subclass
-			throw new Error('Custom provider test not implemented');
-		}
-		throw new Error('No provider configured');
-	}
+	protected abstract sendTestMessage(modelId: string): Promise<any>;
 
 	/**
 	 * Provides the list of available language models for this provider.
@@ -436,11 +421,12 @@ export abstract class ModelProvider implements positron.ai.LanguageModelChatProv
 	 * Provides a chat response from the language model.
 	 *
 	 * This is the main entry point for handling chat interactions with the AI model.
-	 * The method routes to either {@link provideVercelResponse} for Vercel AI SDK
-	 * providers or {@link provideCustomResponse} for custom implementations.
+	 * Subclasses must implement this method to provide chat responses using their
+	 * specific provider SDK.
 	 *
-	 * The response is streamed incrementally through the progress reporter, allowing
-	 * UI to update as tokens are received. Supports both text responses and tool calls.
+	 * The response should be streamed incrementally through the progress reporter,
+	 * allowing UI to update as tokens are received. Should support both text responses
+	 * and tool calls.
 	 *
 	 * @param model - Information about the model to use for the response
 	 * @param messages - Array of chat messages representing the conversation history
@@ -449,336 +435,28 @@ export abstract class ModelProvider implements positron.ai.LanguageModelChatProv
 	 * @param token - Cancellation token to abort the request
 	 * @returns A promise that resolves when the response is complete
 	 *
-	 * @throws {Error} If no provider is configured (neither Vercel nor custom)
-	 *
-	 * @see {@link provideVercelResponse} for Vercel AI SDK implementation
-	 * @see {@link provideCustomResponse} for custom provider implementation
-	 */
-	async provideLanguageModelChatResponse(
-		model: vscode.LanguageModelChatInformation,
-		messages: vscode.LanguageModelChatMessage2[],
-		options: vscode.ProvideLanguageModelChatResponseOptions,
-		progress: vscode.Progress<vscode.LanguageModelResponsePart2>,
-		token: vscode.CancellationToken
-	): Promise<void> {
-		if (this.providerType === 'vercel' && this.aiProvider) {
-			return this.provideVercelResponse(model, messages, options, progress, token);
-		} else if (this.customProvider) {
-			return this.provideCustomResponse(model, messages, options, progress, token);
-		}
-		throw new Error('No provider configured');
-	}
-
-	/**
-	 * Provides a chat response using the Vercel AI SDK.
-	 *
-	 * This is the default implementation for Vercel AI SDK-based providers. It:
-	 * 1. Processes and validates messages
-	 * 2. Converts messages to Vercel AI format
-	 * 3. Sets up tools if provided
-	 * 4. Streams the response using {@link ai.streamText}
-	 * 5. Handles response parts (text, tool calls, errors)
-	 * 6. Tracks token usage
-	 *
-	 * Special handling is included for:
-	 * - Anthropic models: Support for experimental_content in tool results
-	 * - Bedrock models: Cache breakpoint support
-	 * - System prompts: Injected from modelOptions.system
-	 *
-	 * @param model - Information about the model to use
-	 * @param messages - Conversation history to send to the model
-	 * @param options - Generation options including tools and model parameters
-	 * @param progress - Progress reporter for streaming response parts
-	 * @param token - Cancellation token to abort the request
-	 * @returns A promise that resolves when streaming is complete
-	 *
-	 * @throws {Error} If AI provider is not configured
-	 *
-	 * @see {@link handleStreamResponse} for response streaming logic
-	 * @see {@link setupTools} for tool configuration
-	 */
-	protected async provideVercelResponse(
-		model: vscode.LanguageModelChatInformation,
-		messages: vscode.LanguageModelChatMessage2[],
-		options: vscode.ProvideLanguageModelChatResponseOptions,
-		progress: vscode.Progress<vscode.LanguageModelResponsePart2>,
-		token: vscode.CancellationToken
-	): Promise<void> {
-		if (!this.aiProvider) {
-			throw new Error('AI provider not configured');
-		}
-
-		const aiModel = this.aiProvider(model.id);
-		const modelOptions = options.modelOptions ?? {};
-
-		const controller = new AbortController();
-		const signal = controller.signal;
-		token.onCancellationRequested(() => controller.abort());
-
-		let tools: Record<string, ai.Tool> | undefined;
-
-		// Ensure all messages have content
-		const processedMessages = processMessages(messages);
-
-		// Only Anthropic currently supports experimental_content in tool results
-		const toolResultExperimentalContent = this.provider === 'anthropic-api' ||
-			aiModel.modelId.includes('anthropic');
-
-		// Only select Bedrock models support cache breakpoints
-		const bedrockCacheBreakpoint = this.provider === 'amazon-bedrock' &&
-			!aiModel.modelId.includes('anthropic.claude-3-5');
-
-		// Add system prompt from modelOptions.system, if provided
-		if (modelOptions.system) {
-			processedMessages.unshift(new vscode.LanguageModelChatMessage(
-				vscode.LanguageModelChatMessageRole.System,
-				modelOptions.system
-			));
-		}
-
-		// Convert all messages to the Vercel AI format
-		const aiMessages: ai.CoreMessage[] = toAIMessage(
-			processedMessages,
-			toolResultExperimentalContent,
-			bedrockCacheBreakpoint
-		);
-
-		// Set up tools if provided
-		if (options.tools && options.tools.length > 0) {
-			tools = this.setupTools([...options.tools]); // Convert readonly array to mutable
-		}
-
-		const modelTools = this._config.toolCalls ? tools : undefined;
-		const requestId = (options.modelOptions as any)?.requestId;
-
-		this.logger.info(`[vercel] Start request ${requestId} to ${model.name} [${aiModel.modelId}]: ${aiMessages.length} messages`);
-		this.logger.debug(`[${model.name}] SEND ${aiMessages.length} messages, ${modelTools ? Object.keys(modelTools).length : 0} tools`);
-
-		// Stream the response
-		const result = ai.streamText({
-			model: aiModel,
-			messages: aiMessages,
-			maxSteps: modelOptions.maxSteps ?? 50,
-			tools: modelTools,
-			abortSignal: signal,
-			maxTokens: getMaxTokens(aiModel.modelId, 'output', this._config.provider, this._config.maxOutputTokens, this.providerName),
-		});
-
-		await this.handleStreamResponse(result, model, progress, token, requestId);
-	}
-
-	/**
-	 * Provides a chat response using a custom (non-Vercel AI SDK) provider implementation.
-	 *
-	 * Override this method in subclasses that use custom provider implementations
-	 * instead of the Vercel AI SDK. This allows providers to implement their own
-	 * request/response handling logic while still benefiting from the common
-	 * ModelProvider infrastructure.
-	 *
-	 * @param model - Information about the model to use
-	 * @param messages - Conversation history to send to the model
-	 * @param options - Generation options including tools and model parameters
-	 * @param progress - Progress reporter for streaming response parts
-	 * @param token - Cancellation token to abort the request
-	 * @returns A promise that resolves when the response is complete
-	 *
-	 * @throws {Error} Default implementation throws an error prompting to override
-	 *
 	 * @example
 	 * ```typescript
-	 * protected async provideCustomResponse(model, messages, options, progress, token) {
-	 *   const response = await this.customProvider.chat(messages);
-	 *   for (const chunk of response) {
-	 *     progress.report(new vscode.LanguageModelTextPart(chunk));
+	 * async provideLanguageModelChatResponse(model, messages, options, progress, token) {
+	 *   const stream = await this._client.chat.stream({
+	 *     model: model.id,
+	 *     messages: convertMessages(messages),
+	 *   });
+	 *
+	 *   for await (const chunk of stream) {
+	 *     progress.report(new vscode.LanguageModelTextPart(chunk.text));
 	 *   }
 	 * }
 	 * ```
 	 */
-	protected async provideCustomResponse(
+	abstract provideLanguageModelChatResponse(
 		model: vscode.LanguageModelChatInformation,
 		messages: vscode.LanguageModelChatMessage2[],
 		options: vscode.ProvideLanguageModelChatResponseOptions,
 		progress: vscode.Progress<vscode.LanguageModelResponsePart2>,
 		token: vscode.CancellationToken
-	): Promise<void> {
-		throw new Error('Custom provider response not implemented. Override this method in your provider class.');
-	}
+	): Promise<void>;
 
-	/**
-	 * Sets up tools (function calling) for the chat request.
-	 *
-	 * Converts VS Code language model tools to Vercel AI SDK tool format.
-	 * Ensures all tool schemas have proper type information, defaulting to
-	 * 'object' if not specified (required by some providers).
-	 *
-	 * @param tools - Array of VS Code language model chat tools to configure
-	 * @returns A record mapping tool names to Vercel AI SDK tool definitions
-	 *
-	 * @see {@link ai.tool} for Vercel AI SDK tool format
-	 */
-	protected setupTools(tools: vscode.LanguageModelChatTool[]): Record<string, ai.Tool> {
-		return tools.reduce((acc: Record<string, ai.Tool>, tool: vscode.LanguageModelChatTool) => {
-			// Some providers require a type for all tool input schemas
-			const input_schema = tool.inputSchema as Record<string, any> ?? {
-				type: 'object',
-				properties: {},
-				required: [],
-			};
-
-			// Ensure schema has a type field
-			if (!input_schema.type) {
-				this.logger.warn(`Tool '${tool.name}' is missing input schema type; defaulting to 'object'`);
-				input_schema.type = 'object';
-			}
-
-			acc[tool.name] = ai.tool({
-				description: tool.description,
-				parameters: ai.jsonSchema(input_schema),
-			});
-			return acc;
-		}, {});
-	}
-
-	/**
-	 * Handles the streaming response from the AI model.
-	 *
-	 * This method processes the stream from {@link ai.streamText}, handling different
-	 * part types (text, reasoning, tool calls, errors) and reporting them through
-	 * the progress reporter. It also:
-	 * - Accumulates text deltas for more efficient logging
-	 * - Flushes accumulated deltas when non-text parts are received
-	 * - Handles warnings from the model
-	 * - Tracks and reports token usage
-	 * - Respects cancellation tokens
-	 *
-	 * @param result - The streaming result from {@link ai.streamText}
-	 * @param model - Information about the model being used
-	 * @param progress - Progress reporter for sending response parts to the caller
-	 * @param token - Cancellation token to abort streaming
-	 * @param requestId - Optional request ID for tracking and logging
-	 * @returns A promise that resolves when streaming is complete
-	 *
-	 * @throws {Error} If an error part is received in the stream
-	 *
-	 * @see {@link handleTokenUsage} for token usage tracking
-	 */
-	protected async handleStreamResponse(
-		result: any,
-		model: vscode.LanguageModelChatInformation,
-		progress: vscode.Progress<vscode.LanguageModelResponsePart2>,
-		token: vscode.CancellationToken,
-		requestId?: string
-	): Promise<void> {
-		let accumulatedTextDeltas: string[] = [];
-
-		const flushAccumulatedTextDeltas = () => {
-			if (accumulatedTextDeltas.length > 0) {
-				const combinedText = accumulatedTextDeltas.join('');
-				this.logger.trace(`[${model.name}] RECV text-delta (${accumulatedTextDeltas.length} parts): ${combinedText}`);
-				accumulatedTextDeltas = [];
-			}
-		};
-
-		for await (const part of result.fullStream) {
-			if (token.isCancellationRequested) {
-				break;
-			}
-
-			if (part.type === 'reasoning') {
-				flushAccumulatedTextDeltas();
-				this.logger.trace(`[${this._config.name}] RECV reasoning: ${part.textDelta}`);
-				progress.report(new vscode.LanguageModelTextPart(part.textDelta));
-			}
-
-			if (part.type === 'text-delta') {
-				accumulatedTextDeltas.push(part.textDelta);
-				progress.report(new vscode.LanguageModelTextPart(part.textDelta));
-			}
-
-			if (part.type === 'tool-call') {
-				flushAccumulatedTextDeltas();
-				this.logger.trace(`[${this._config.name}] RECV tool-call: ${part.toolCallId} (${part.toolName}) with args: ${JSON.stringify(part.args)}`);
-				progress.report(new vscode.LanguageModelToolCallPart(part.toolCallId, part.toolName, part.args));
-			}
-
-			if (part.type === 'error') {
-				flushAccumulatedTextDeltas();
-				this.logger.warn(`[${model.name}] RECV error`, part.error);
-				const errorMsg = await this.parseProviderError(part.error) ||
-					(typeof part.error === 'string' ? part.error : JSON.stringify(part.error, null, 2));
-				throw new Error(`[${model.name}] Error in chat response: ${errorMsg}`);
-			}
-		}
-
-		// Flush any remaining accumulated text deltas
-		flushAccumulatedTextDeltas();
-
-		// Log warnings
-		const warnings = await result.warnings;
-		if (warnings) {
-			for (const warning of warnings) {
-				this.logger.warn(`[${model.id}] ${warning}`);
-			}
-		}
-
-		// Handle token usage
-		await this.handleTokenUsage(result, model, requestId);
-	}
-
-	/**
-	 * Handles token usage tracking and reporting.
-	 *
-	 * Extracts token usage information from the AI result and:
-	 * - Tracks input, output, and cached tokens
-	 * - Handles provider-specific usage metadata (e.g., Bedrock cache tokens)
-	 * - Records usage in request tracking and extension storage
-	 * - Logs usage information for debugging
-	 *
-	 * @param result - The AI result containing usage information
-	 * @param model - Information about the model that was used
-	 * @param requestId - Optional request ID for tracking this specific request
-	 * @returns A promise that resolves when usage tracking is complete
-	 *
-	 * @see {@link recordRequestTokenUsage} for request-specific tracking
-	 * @see {@link recordTokenUsage} for persistent storage
-	 */
-	protected async handleTokenUsage(
-		result: any,
-		model: vscode.LanguageModelChatInformation,
-		requestId?: string
-	): Promise<void> {
-		const usage = await result.usage;
-		const metadata = await result.providerMetadata;
-		const tokens: TokenUsage = {
-			inputTokens: usage.promptTokens,
-			outputTokens: usage.completionTokens,
-			cachedTokens: 0,
-			providerMetadata: metadata,
-		};
-
-		// Handle Bedrock-specific usage
-		if (metadata && metadata.bedrock && metadata.bedrock.usage) {
-			const metaUsage = metadata.bedrock.usage as Record<string, any>;
-			tokens.inputTokens += metaUsage.cacheWriteInputTokens || 0;
-			tokens.cachedTokens += metaUsage.cacheReadInputTokens || 0;
-
-			// Report token usage information
-			const part: any = vscode.LanguageModelDataPart.json({ type: 'usage', data: tokens });
-			(part as any).report && (part as any).report(part);
-
-			this.logger.debug(`[${model.name}]: Bedrock usage: ${JSON.stringify(usage, null, 2)}`);
-		}
-
-		if (requestId) {
-			recordRequestTokenUsage(requestId, this.provider, tokens);
-		}
-
-		if (this._context) {
-			recordTokenUsage(this._context, this.provider, tokens);
-		}
-
-		this.logger.info(`[vercel]: End request ${requestId}; usage: ${tokens.inputTokens} input tokens (+${tokens.cachedTokens} cached), ${tokens.outputTokens} output tokens`);
-	}
 
 	/**
 	 * Provides an estimated token count for the given text or messages.
@@ -933,16 +611,12 @@ export abstract class ModelProvider implements positron.ai.LanguageModelChatProv
 
 		this.logger.info(`Using ${configuredModels.length} configured models.`);
 
-		if (!this.aiProvider) {
-			this.aiProvider = this.createAIProvider();
-		}
-
 		const models: vscode.LanguageModelChatInformation[] = configuredModels.map(model =>
 			createModelInfo({
 				id: model.identifier,
 				name: model.name,
 				family: this.provider,
-				version: this.aiProvider ? this.aiProvider(model.identifier).specificationVersion : '1.0',
+				version: '1.0',
 				provider: this.provider,
 				providerName: this.providerName,
 				capabilities: this.capabilities,
@@ -998,27 +672,6 @@ export abstract class ModelProvider implements positron.ai.LanguageModelChatProv
 	protected createDefaultModel(): vscode.LanguageModelChatInformation[] {
 		this.logger.info('No models available; returning default model information.');
 
-		if (!this.aiProvider) {
-			this.aiProvider = this.createAIProvider();
-		}
-
-		if (this.aiProvider) {
-			const aiModel = this.aiProvider(this._config.model, this.aiOptions);
-			const modelInfo = createModelInfo({
-				id: aiModel.modelId,
-				name: this.name,
-				family: aiModel.provider,
-				version: aiModel.specificationVersion,
-				provider: this._config.provider,
-				providerName: this.providerName,
-				capabilities: this.capabilities,
-				defaultMaxInput: this._config.maxInputTokens,
-				defaultMaxOutput: this._config.maxOutputTokens
-			});
-			return [{ ...modelInfo, isDefault: true }];
-		}
-
-		// For custom providers
 		const modelInfo = createModelInfo({
 			id: this._config.model,
 			name: this.name,
