+++ extensions/positron-assistant/src/providers/base/modelProviderTypes.ts
@@ -5,6 +5,13 @@
 
 import * as vscode from 'vscode';
 import * as positron from 'positron';
+import * as ai from 'ai';
+
+/**
+ * Type for Vercel AI SDK provider factory functions.
+ * Creates a language model instance given a model ID and optional configuration.
+ */
+export type AIProviderFactory = (id: string, options?: Record<string, any>) => ai.LanguageModelV1;
 
 /**
  * Common types and interfaces for model providers.
@@ -16,14 +23,6 @@ import * as positron from 'positron';
  *
  * @module modelProviderTypes
  */
-
-/**
- * Type of provider implementation.
- *
- * - `vercel`: Provider uses the Vercel AI SDK for model interactions
- * - `custom`: Provider uses a custom implementation without Vercel AI SDK
- */
-export type ProviderType = 'vercel' | 'custom';
 
 /**
  * Model capabilities that can be supported by a provider.
