+++ extensions/positron-assistant/src/providers/mistral/mistralProvider.ts
@@ -6,14 +6,15 @@
 import * as vscode from 'vscode';
 import * as positron from 'positron';
 import { createMistral, MistralProvider } from '@ai-sdk/mistral';
-import { ModelProvider } from '../base/modelProvider';
+import { VercelModelProvider } from '../base/vercelModelProvider';
+import { AIProviderFactory } from '../base/modelProviderTypes';
 import { ModelConfig } from '../../config';
 
 /**
  * Mistral AI model provider implementation.
  * Supports Mistral's models via the Mistral API.
  */
-export class MistralLanguageModel extends ModelProvider implements positron.ai.LanguageModelChatProvider {
+export class MistralLanguageModel extends VercelModelProvider implements positron.ai.LanguageModelChatProvider {
 	protected declare aiProvider: MistralProvider;
 
 	static source: positron.ai.LanguageModelSource = {
@@ -34,8 +35,6 @@ export class MistralLanguageModel extends ModelProvider implements positron.ai.L
 
 	constructor(_config: ModelConfig, _context?: vscode.ExtensionContext) {
 		super(_config, _context);
-		this.initializeLogger();
-		this.initializeProvider();
 	}
 
 	/**
@@ -52,7 +51,7 @@ export class MistralLanguageModel extends ModelProvider implements positron.ai.L
 	 * Creates the AI provider instance.
 	 * @returns The Mistral provider function.
 	 */
-	protected createAIProvider(): any {
+	protected createAIProvider(): AIProviderFactory | undefined {
 		return this.aiProvider;
 	}
